pyproject.toml
[tool.poetry]
name = "magnetic_one_langgraph"
version = "0.1.0"
description = "LangGraph-based agent system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "2.9.2"
pydantic-settings = "^2.0.0"
langgraph = "0.2.48"
langchain = "0.3.7"
langchain-core = "0.3.19"
langchain-openai = "0.2.8"
python-dotenv = "1.0.1"
tavily-python = "0.5.0"
numpy = "1.26.4"
openai = "1.54.4"
tiktoken = "0.8.0"
httpx = "0.27.2"
aiohttp = "3.11.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

